requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==24.2.0
backports.tarfile==1.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
distlib==0.3.9
docutils==0.21.2
fastapi==0.115.8
filelock==3.17.0
Flask==3.1.0
google-auth==2.38.0
google-genai==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
Jinja2==3.1.6
JsonForm==0.0.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
JsonSir==0.0.2
keyring==25.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.4.0
nh3==0.2.18
pkginfo==1.10.0
platformdirs==4.3.6
podman==5.4.0.1
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
Python-EasyConfig==0.1.7
python-multipart==0.0.20
pywin32-ctypes==0.2.3
PyYAML==6.0.2
readme_renderer==44.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
Resource==0.2.1
rfc3986==2.0.0
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.45.3
twine==5.1.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.34.0
virtualenv==20.29.1
websockets==14.2
Werkzeug==3.1.3
zipp==3.20.1
gunicorn
python-dotenv